paste_secure/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
Markdown==3.5.2
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4